requirements.txt
Django==5.1.3
djangorestframework==3.14.0
djoser==2.1.0
django-filter==23.1
celery==5.3.0
redis==4.5.0
